backend/tests/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database (PostgreSQL via SQLAlchemy)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
redis==5.0.1
aiofiles==23.2.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring and Metrics (for tests)
prometheus_client==0.19.0

# Machine Learning (for FAQ clustering tests)
scikit-learn==1.3.2
numpy==1.24.4

# Additional testing dependencies
asyncio-mqtt==0.13.0
mock==5.1.0
